Move some data and shaping algorithms into `parley_core`
