+++ parley/src/layout/data.rs
@@ -2,117 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 use crate::inline_box::InlineBox;
-use crate::layout::{ContentWidths, Glyph, LineMetrics, RunMetrics, Style};
+use crate::layout::{ContentWidths, LineMetrics, RunMetrics, Style};
 use crate::style::Brush;
 use crate::util::nearly_zero;
 use crate::{FontData, IndentOptions, InlineBoxKind, LineHeight, OverflowWrap, TextWrapMode};
 use core::ops::Range;
 
 use alloc::vec::Vec;
-use parley_core::{Boundary, CharInfo};
-
-use crate::analysis::cluster::Whitespace;
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub(crate) struct ClusterData {
-    pub(crate) info: ClusterInfo,
-    /// Cluster flags (see impl methods for details).
-    pub(crate) flags: u16,
-    /// Style index for this cluster.
-    pub(crate) style_index: u16,
-    /// Number of glyphs in this cluster (0xFF = single glyph stored inline)
-    pub(crate) glyph_len: u8,
-    /// Number of text bytes in this cluster
-    pub(crate) text_len: u8,
-    /// If `glyph_len == 0xFF`, then `glyph_offset` is a glyph identifier,
-    /// otherwise, it's an offset into the glyph array with the base
-    /// taken from the owning run.
-    pub(crate) glyph_offset: u32,
-    /// Offset into the text for this cluster
-    pub(crate) text_offset: u16,
-    /// Advance width for this cluster
-    pub(crate) advance: f32,
-}
-
-impl ClusterData {
-    pub(crate) const LIGATURE_START: u16 = 1;
-    pub(crate) const LIGATURE_COMPONENT: u16 = 2;
-
-    #[inline(always)]
-    pub(crate) fn is_ligature_start(self) -> bool {
-        self.flags & Self::LIGATURE_START != 0
-    }
-
-    #[inline(always)]
-    pub(crate) fn is_ligature_component(self) -> bool {
-        self.flags & Self::LIGATURE_COMPONENT != 0
-    }
-
-    #[inline(always)]
-    pub(crate) fn text_range(self, run: &RunData) -> Range<usize> {
-        let start = run.text_range.start + self.text_offset as usize;
-        start..start + self.text_len as usize
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub(crate) struct ClusterInfo {
-    boundary: Boundary,
-    source_char: char,
-}
-
-impl ClusterInfo {
-    pub(crate) fn new(boundary: Boundary, source_char: char) -> Self {
-        Self {
-            boundary,
-            source_char,
-        }
-    }
-
-    // Returns the boundary type of the cluster.
-    pub(crate) fn boundary(self) -> Boundary {
-        self.boundary
-    }
-
-    // Returns the whitespace type of the cluster.
-    pub(crate) fn whitespace(self) -> Whitespace {
-        to_whitespace(self.source_char)
-    }
-
-    /// Returns if the cluster is a line boundary.
-    pub(crate) fn is_boundary(self) -> bool {
-        self.boundary != Boundary::None
-    }
-
-    /// Returns if the cluster is an emoji.
-    pub(crate) fn is_emoji(self) -> bool {
-        // TODO: Defer to ICU4X properties (see: https://docs.rs/icu/latest/icu/properties/props/struct.Emoji.html).
-        matches!(self.source_char as u32, 0x1F600..=0x1F64F | 0x1F300..=0x1F5FF | 0x1F680..=0x1F6FF | 0x2600..=0x26FF | 0x2700..=0x27BF)
-    }
-
-    /// Returns if the cluster is any whitespace.
-    pub(crate) fn is_whitespace(self) -> bool {
-        self.source_char.is_whitespace()
-    }
-
-    /// Returns the cluster's original character.
-    pub(crate) fn source_char(self) -> char {
-        self.source_char
-    }
-}
-
-const fn to_whitespace(c: char) -> Whitespace {
-    const LINE_SEPARATOR: char = '\u{2028}';
-    const PARAGRAPH_SEPARATOR: char = '\u{2029}';
-
-    match c {
-        ' ' => Whitespace::Space,
-        '\t' => Whitespace::Tab,
-        '\n' | '\r' | LINE_SEPARATOR | PARAGRAPH_SEPARATOR => Whitespace::Newline,
-        '\u{00A0}' => Whitespace::NoBreakSpace,
-        _ => Whitespace::None,
-    }
-}
+use parley_core::shape::{ClusterData, ClusterInfo, Whitespace, to_whitespace};
+use parley_core::{Boundary, CharInfo, Glyph};
 
 /// `HarfRust`-based run data
 #[derive(Clone, Debug, PartialEq)]
