+++ parley/src/layout/mod.rs
@@ -7,7 +7,6 @@
 mod accessibility;
 mod alignment;
 mod cluster;
-mod glyph;
 mod line;
 mod line_break;
 mod run;
@@ -26,7 +25,6 @@ pub use accessibility::LayoutAccessibility;
 pub use alignment::{Alignment, AlignmentOptions};
 pub use cluster::{Affinity, Cluster, ClusterPath, ClusterSide};
 pub use data::BreakReason;
-pub use glyph::Glyph;
 pub use layout::Layout;
 pub use line::{GlyphRun, Line, LineMetrics, PositionedInlineBox, PositionedLayoutItem};
 pub use line_break::{
